+++ src/combinator/provided/till.rs
@@ -1,50 +1,71 @@
-use crate::{combinator::wrap, C};
+use crate::{
+  action::{Input, Output},
+  combinator::wrap,
+  C,
+};
 
 /// A util trait to make [`till`] generic over different types.
 ///
 /// Built-in implementations are provided for [`String`], `&str`, [`char`] and `()`.
 ///
 /// See [`till`] for more details.
-/// # Safety
-/// You should ensure that [`Output::rest`](crate::combinator::Output::rest) can be built
-/// as a valid UTF-8 string.
-/// This will be checked using [`debug_assert!`].
-pub unsafe trait Till {
-  /// Check if the input contains this instance.
-  /// Return the length of digested bytes if found.
-  fn parse(&self, input: &str) -> Option<usize>;
+pub trait Till<State, Heap> {
+  /// Check if the rest of input text contains this instance.
+  /// Return the output after digesting the instance if found.
+  fn exec<'text>(
+    &self,
+    input: &mut Input<'text, &mut State, &mut Heap>,
+  ) -> Option<Output<'text, ()>>;
 }
 
-unsafe impl Till for String {
+impl<State, Heap> Till<State, Heap> for &str {
   #[inline]
-  fn parse(&self, input: &str) -> Option<usize> {
-    input.find(self).map(|i| i + self.len())
+  fn exec<'text>(
+    &self,
+    input: &mut Input<'text, &mut State, &mut Heap>,
+  ) -> Option<Output<'text, ()>> {
+    input
+      .rest()
+      .find(self)
+      .map(|i| unsafe { input.digest_unchecked(i + self.len()) })
   }
 }
 
-unsafe impl Till for &str {
+impl<State, Heap> Till<State, Heap> for String {
   #[inline]
-  fn parse(&self, input: &str) -> Option<usize> {
-    input.find(self).map(|i| i + self.len())
+  fn exec<'text>(
+    &self,
+    input: &mut Input<'text, &mut State, &mut Heap>,
+  ) -> Option<Output<'text, ()>> {
+    self.as_str().exec(input)
   }
 }
 
-unsafe impl Till for char {
+impl<State, Heap> Till<State, Heap> for char {
   #[inline]
-  fn parse(&self, input: &str) -> Option<usize> {
-    input.find(*self).map(|i| i + self.len_utf8())
+  fn exec<'text>(
+    &self,
+    input: &mut Input<'text, &mut State, &mut Heap>,
+  ) -> Option<Output<'text, ()>> {
+    input
+      .rest()
+      .find(*self)
+      .map(|i| unsafe { input.digest_unchecked(i + self.len_utf8()) })
   }
 }
 
-unsafe impl Till for () {
+impl<State, Heap> Till<State, Heap> for () {
   #[inline]
-  fn parse(&self, input: &str) -> Option<usize> {
-    Some(input.len())
+  fn exec<'text>(&self, _: &mut Input<'text, &mut State, &mut Heap>) -> Option<Output<'text, ()>> {
+    Some(Output {
+      value: (),
+      rest: "",
+    })
   }
 }
 
 /// Return a combinator to match the provided pattern, eat all the bytes
-/// to the end of the first occurrence of the pattern.
+/// to the end of the first occurrence of the pattern (inclusive).
 ///
 /// Empty string is allowed, but be careful with infinite loops.
 /// # Examples
@@ -65,12 +86,8 @@ unsafe impl Till for () {
 /// # );
 /// ```
 #[inline]
-pub const fn till<State, Heap>(pattern: impl Till) -> C!((), State, Heap) {
-  wrap(move |input| {
-    pattern
-      .parse(input.rest())
-      .map(|digested| unsafe { input.digest_unchecked(digested) })
-  })
+pub const fn till<State, Heap>(pattern: impl Till<State, Heap>) -> C!((), State, Heap) {
+  wrap(move |input| pattern.exec(input))
 }
 
 #[cfg(test)]
@@ -79,7 +96,7 @@ mod tests {
   use crate::action::{Action, Input, Output};
 
   #[test]
-  fn till_parse() {
+  fn until_exec() {
     assert_eq!(
       till("end".to_string()).exec(&mut Input::new("123end456", 0, &mut (), &mut ()).unwrap()),
       Some(Output {